Change Request#path to Request#path_info.
In some versions of Rack, at least, Request#path is called #fullpath, but #path_info should be consistent and is probably more correct anyways, as it will scope it to the path the app was mounted to. Otherwise if, for whatever reason, you mounted the app below the root you would not be able to access any of the __*__ commands at all.